Identifying missed graphics
Now leaves a message in the output raw file where graphics are found in
the source raw file but have no equivalent in the source graphics file.
Making it easy to see what modded-in entities may need your attention to
have consistent graphics. Search for BAMM in the output files.